LPD-72676 Refactor conversion logic out into separate DTOConverter classes
