+++ modules/apps/headless/headless-admin-site/headless-admin-site-impl/src/main/java/com/liferay/headless/admin/site/internal/resource/v1_0/NavigationMenuResourceImpl.java
@@ -6,53 +6,34 @@
 package com.liferay.headless.admin.site.internal.resource.v1_0;
 
 import com.liferay.batch.engine.thread.local.BatchEngineThreadLocal;
-import com.liferay.document.library.kernel.model.DLFileEntry;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.exportimport.vulcan.batch.engine.ExportImportVulcanBatchEngineTaskItemDelegate;
-import com.liferay.headless.admin.site.dto.v1_0.DisplayPageNavigationMenuItemSettings;
 import com.liferay.headless.admin.site.dto.v1_0.NavigationMenu;
 import com.liferay.headless.admin.site.dto.v1_0.NavigationMenuItem;
-import com.liferay.headless.admin.site.dto.v1_0.PageNavigationMenuItemSettings;
-import com.liferay.headless.admin.site.dto.v1_0.UrlNavigationMenuItemSettings;
-import com.liferay.headless.admin.site.dto.v1_0.VocabularyNavigationMenuItemSettings;
 import com.liferay.headless.admin.site.internal.odata.entity.v1_0.NavigationMenuEntityModel;
 import com.liferay.headless.admin.site.internal.resource.v1_0.util.GroupUtil;
 import com.liferay.headless.admin.site.resource.v1_0.NavigationMenuResource;
-import com.liferay.headless.admin.user.dto.v1_0.Creator;
 import com.liferay.headless.common.spi.service.context.ServiceContextBuilder;
-import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
-import com.liferay.portal.kernel.json.JSONException;
 import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.Layout;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.Sort;
 import com.liferay.portal.kernel.search.filter.Filter;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.LayoutLocalService;
-import com.liferay.portal.kernel.service.PermissionService;
-import com.liferay.portal.kernel.service.ResourceActionLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.kernel.util.UnicodePropertiesBuilder;
 import com.liferay.portal.odata.entity.EntityModel;
 import com.liferay.portal.vulcan.custom.field.CustomFieldsUtil;
-import com.liferay.portal.vulcan.fields.NestedFieldsSupplier;
+import com.liferay.portal.vulcan.dto.converter.DTOConverter;
+import com.liferay.portal.vulcan.dto.converter.DTOConverterRegistry;
+import com.liferay.portal.vulcan.dto.converter.DefaultDTOConverterContext;
 import com.liferay.portal.vulcan.pagination.Page;
 import com.liferay.portal.vulcan.pagination.Pagination;
-import com.liferay.portal.vulcan.permission.Permission;
-import com.liferay.portal.vulcan.permission.PermissionUtil;
-import com.liferay.portal.vulcan.util.LocalizedMapUtil;
 import com.liferay.portal.vulcan.util.SearchUtil;
 import com.liferay.site.navigation.admin.constants.SiteNavigationAdminPortletKeys;
 import com.liferay.site.navigation.constants.SiteNavigationActionKeys;
@@ -65,18 +46,14 @@ import com.liferay.site.navigation.service.SiteNavigationMenuLocalService;
 import com.liferay.site.navigation.service.SiteNavigationMenuService;
 import com.liferay.site.navigation.type.SiteNavigationMenuItemType;
 import com.liferay.site.navigation.type.SiteNavigationMenuItemTypeRegistry;
-import com.liferay.site.navigation.util.comparator.SiteNavigationMenuItemOrderComparator;
 
 import jakarta.ws.rs.core.MultivaluedMap;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -327,8 +304,7 @@ public class NavigationMenuResourceImpl
 		if (!_hasModel(groupId, navigationMenuItem, unicodeProperties)) {
 			throw new IllegalArgumentException(
 				"Unable to find navigation menu item with external reference " +
-					"code " +
-						navigationMenuItem.getExternalReferenceCode());
+					"code " + navigationMenuItem.getExternalReferenceCode());
 		}
 
 		SiteNavigationMenuItem siteNavigationMenuItem =
@@ -381,11 +357,6 @@ public class NavigationMenuResourceImpl
 		}
 	}
 
-	private Locale _getLocaleFromProperty(Map.Entry<String, String> property) {
-		return LocaleUtil.fromLanguageId(
-			StringUtil.removeSubstring(property.getKey(), "name_"));
-	}
-
 	private String _getLocalizedNamesFromI18nMap(
 		NavigationMenuItem navigationMenuItem) {
 
@@ -408,42 +379,6 @@ public class NavigationMenuResourceImpl
 		}
 
 		return localizedNamesJSONObject.toString();
-	}
-
-	private Map<Locale, String> _getLocalizedNamesFromProperties(
-			UnicodeProperties unicodeProperties)
-		throws JSONException {
-
-		if (unicodeProperties == null) {
-			return new HashMap<>();
-		}
-
-		String localizedNames = unicodeProperties.getProperty("localizedNames");
-
-		Map<Locale, String> properties = new HashMap<>();
-
-		if (localizedNames != null) {
-			JSONObject localizedNamesJSONObject = _jsonFactory.createJSONObject(
-				localizedNames);
-
-			for (String key : localizedNamesJSONObject.keySet()) {
-				properties.put(
-					LocaleUtil.fromLanguageId(key),
-					localizedNamesJSONObject.getString(key));
-			}
-
-			return properties;
-		}
-
-		for (Map.Entry<String, String> entry : unicodeProperties.entrySet()) {
-			if (!_isNameProperty(entry)) {
-				continue;
-			}
-
-			properties.put(_getLocaleFromProperty(entry), entry.getValue());
-		}
-
-		return properties;
 	}
 
 	private Map<String, String> _getLocalizedPropertiesMapFromI18nMap(
@@ -470,136 +405,6 @@ public class NavigationMenuResourceImpl
 		return localizedPropertyMap;
 	}
 
-	private String _getName(
-			Layout layout, String type, UnicodeProperties unicodeProperties,
-			boolean useCustomName)
-		throws JSONException {
-
-		String defaultLanguageId = LocaleUtil.toLanguageId(
-			LocaleUtil.getDefault());
-		String preferredLanguageId =
-			contextAcceptLanguage.getPreferredLanguageId();
-
-		if (StringUtil.equals(type, "page")) {
-			if (!useCustomName && (layout == null)) {
-				return null;
-			}
-
-			if (!useCustomName && (layout != null)) {
-				return layout.getName(
-					contextAcceptLanguage.getPreferredLocale());
-			}
-
-			if (!useCustomName) {
-				return null;
-			}
-
-			return unicodeProperties.getProperty(
-				"name_" + preferredLanguageId,
-				unicodeProperties.getProperty("name_" + defaultLanguageId));
-		}
-
-		if (useCustomName) {
-			JSONObject customNameJSONObject = _jsonFactory.createJSONObject(
-				unicodeProperties.getProperty("localizedNames"));
-
-			return customNameJSONObject.getString(
-				preferredLanguageId,
-				customNameJSONObject.getString(defaultLanguageId));
-		}
-
-		if (StringUtil.equals(type, "navigationMenu") ||
-			StringUtil.equals(type, "url")) {
-
-			return unicodeProperties.getProperty(
-				"name_" + preferredLanguageId,
-				unicodeProperties.getProperty("name_" + defaultLanguageId));
-		}
-
-		return unicodeProperties.getProperty("title");
-	}
-
-	private Object _getNavigationMenuItemSettings(
-		String type, UnicodeProperties unicodeProperties) {
-
-		if (Objects.equals(
-				type, SiteNavigationMenuItemTypeConstants.ASSET_VOCABULARY)) {
-
-			return new VocabularyNavigationMenuItemSettings() {
-				{
-					setClassName(
-						() -> unicodeProperties.getProperty("className"));
-					setExternalReferenceCode(
-						() -> unicodeProperties.getProperty(
-							"externalReferenceCode"));
-					setScopeExternalReferenceCode(
-						() -> unicodeProperties.getProperty(
-							"scopeExternalReferenceCode"));
-					setShowAssetVocabularyLevel(
-						() -> Boolean.valueOf(
-							unicodeProperties.getProperty(
-								"showAssetVocabularyLevel")));
-					setTitle(() -> unicodeProperties.getProperty("title"));
-					setType(() -> unicodeProperties.getProperty("type"));
-				}
-			};
-		}
-		else if (Objects.equals(
-					type, SiteNavigationMenuItemTypeConstants.LAYOUT)) {
-
-			return new PageNavigationMenuItemSettings() {
-				{
-					setExternalReferenceCode(
-						() -> unicodeProperties.getProperty(
-							"externalReferenceCode"));
-					setPrivatePage(
-						() -> Boolean.valueOf(
-							unicodeProperties.getProperty("privateLayout")));
-				}
-			};
-		}
-		else if (Objects.equals(
-					type, SiteNavigationMenuItemTypeConstants.URL)) {
-
-			return new UrlNavigationMenuItemSettings() {
-				{
-					setUrl(() -> unicodeProperties.getProperty("url"));
-					setUseNewTab(
-						() -> Boolean.valueOf(
-							unicodeProperties.getProperty("useNewTab")));
-				}
-			};
-		}
-
-		SiteNavigationMenuItemType siteNavigationMenuItemType =
-			_siteNavigationMenuItemTypeRegistry.getSiteNavigationMenuItemType(
-				type);
-
-		Class<?> clazz = siteNavigationMenuItemType.getClass();
-
-		if (Objects.equals(
-				clazz.getName(),
-				SiteNavigationMenuItemTypeConstants.DISPLAY_PAGE_CLASS_NAME)) {
-
-			return new DisplayPageNavigationMenuItemSettings() {
-				{
-					setClassName(
-						() -> unicodeProperties.getProperty("className"));
-					setExternalReferenceCode(
-						() -> unicodeProperties.getProperty(
-							"externalReferenceCode"));
-					setScopeExternalReferenceCode(
-						() -> unicodeProperties.getProperty(
-							"scopeExternalReferenceCode"));
-					setTitle(() -> unicodeProperties.getProperty("title"));
-					setType(() -> unicodeProperties.getProperty("type"));
-				}
-			};
-		}
-
-		return null;
-	}
-
 	private int _getNavigationMenuType(NavigationMenu navigationMenu) {
 		int type = SiteNavigationConstants.TYPE_DEFAULT;
 
@@ -619,50 +424,6 @@ public class NavigationMenuResourceImpl
 		}
 
 		return null;
-	}
-
-	private Map<Long, List<SiteNavigationMenuItem>>
-		_getSiteNavigationMenuItemsMap(
-			List<SiteNavigationMenuItem> siteNavigationMenuItems) {
-
-		Map<Long, List<SiteNavigationMenuItem>> siteNavigationMenuItemsMap =
-			new HashMap<>();
-
-		for (SiteNavigationMenuItem siteNavigationMenuItem :
-				siteNavigationMenuItems) {
-
-			long parentSiteNavigationMenuItemId =
-				siteNavigationMenuItem.getParentSiteNavigationMenuItemId();
-
-			if (siteNavigationMenuItemsMap.containsKey(
-					parentSiteNavigationMenuItemId)) {
-
-				continue;
-			}
-
-			for (SiteNavigationMenuItem childSiteNavigationMenuItem :
-					siteNavigationMenuItems) {
-
-				if (parentSiteNavigationMenuItemId !=
-						childSiteNavigationMenuItem.
-							getParentSiteNavigationMenuItemId()) {
-
-					continue;
-				}
-
-				List<SiteNavigationMenuItem> parentSiteNavigationMenuItems =
-					siteNavigationMenuItemsMap.getOrDefault(
-						parentSiteNavigationMenuItemId, new ArrayList<>());
-
-				parentSiteNavigationMenuItems.add(childSiteNavigationMenuItem);
-
-				siteNavigationMenuItemsMap.put(
-					parentSiteNavigationMenuItemId,
-					parentSiteNavigationMenuItems);
-			}
-		}
-
-		return siteNavigationMenuItemsMap;
 	}
 
 	private String _getTypeSettings(NavigationMenuItem navigationMenuItem) {
@@ -855,18 +616,6 @@ public class NavigationMenuResourceImpl
 		return unicodeProperties.toString();
 	}
 
-	private UnicodeProperties _getUnicodeProperties(
-		SiteNavigationMenuItem siteNavigationMenuItem) {
-
-		if (siteNavigationMenuItem == null) {
-			return null;
-		}
-
-		return UnicodePropertiesBuilder.fastLoad(
-			siteNavigationMenuItem.getTypeSettings()
-		).build();
-	}
-
 	private boolean _hasModel(
 			long groupId, NavigationMenuItem navigationMenuItem,
 			UnicodeProperties unicodeProperties)
@@ -894,234 +643,29 @@ public class NavigationMenuResourceImpl
 		return auto;
 	}
 
-	private boolean _isNameProperty(Map.Entry<String, String> entry) {
-		String key = entry.getKey();
-
-		return key.startsWith("name_");
-	}
-
 	private NavigationMenu _toNavigationMenu(
-		SiteNavigationMenu siteNavigationMenu) {
-
-		return new NavigationMenu() {
-			{
-				setActions(
-					() -> HashMapBuilder.put(
-						"delete",
-						addAction(
-							ActionKeys.DELETE, siteNavigationMenu,
-							"deleteSiteNavigationMenu")
-					).put(
-						"replace",
-						addAction(
-							ActionKeys.UPDATE, siteNavigationMenu,
-							"putSiteNavigationMenu")
-					).build());
-				setAuto(siteNavigationMenu::getAuto);
-				setCreator(
-					() -> {
-						User user = _userLocalService.fetchUser(
-							siteNavigationMenu.getUserId());
-
-						if (user == null) {
-							return null;
-						}
-
-						return new Creator() {
-							{
-								setExternalReferenceCode(
-									user::getExternalReferenceCode);
-							}
-						};
-					});
-				setDateCreated(siteNavigationMenu::getCreateDate);
-				setDateModified(siteNavigationMenu::getModifiedDate);
-				setExternalReferenceCode(
-					siteNavigationMenu::getExternalReferenceCode);
-				setId(siteNavigationMenu::getSiteNavigationMenuId);
-				setName(siteNavigationMenu::getName);
-				setNavigationMenuItems(
-					() -> {
-						Map<Long, List<SiteNavigationMenuItem>>
-							siteNavigationMenuItemsMap =
-								_getSiteNavigationMenuItemsMap(
-									_siteNavigationMenuItemService.
-										getSiteNavigationMenuItems(
-											siteNavigationMenu.
-												getSiteNavigationMenuId(),
-											SiteNavigationMenuItemOrderComparator.
-												getInstance(true)));
-
-						return transformToArray(
-							siteNavigationMenuItemsMap.getOrDefault(
-								0L, new ArrayList<>()),
-							siteNavigationMenuItem -> _toNavigationMenuItem(
-								siteNavigationMenuItem,
-								siteNavigationMenuItemsMap),
-							NavigationMenuItem.class);
-					});
-				setNavigationType(
-					() -> {
-						if (siteNavigationMenu.getType() == 0) {
-							return null;
-						}
-
-						return NavigationType.values()
-							[siteNavigationMenu.getType() - 1];
-					});
-				setPermissions(() -> _toPermissions(siteNavigationMenu));
-				setSiteExternalReferenceCode(
-					() -> {
-						Group group = _groupLocalService.fetchGroup(
-							siteNavigationMenu.getGroupId());
-
-						if (group != null) {
-							return group.getExternalReferenceCode();
-						}
-
-						return StringPool.BLANK;
-					});
-			}
-		};
-	}
-
-	private NavigationMenuItem _toNavigationMenuItem(
-		SiteNavigationMenuItem siteNavigationMenuItem,
-		Map<Long, List<SiteNavigationMenuItem>> siteNavigationMenuItemsMap) {
-
-		UnicodeProperties unicodeProperties = _getUnicodeProperties(
-			siteNavigationMenuItem);
-
-		Layout layout = _layoutLocalService.fetchLayoutByExternalReferenceCode(
-			unicodeProperties.getProperty("externalReferenceCode"),
-			siteNavigationMenuItem.getGroupId());
-
-		String navigationMenuItemType = _toType(
-			siteNavigationMenuItem.getType());
-
-		return new NavigationMenuItem() {
-			{
-				setAvailableLanguages(
-					() -> {
-						Map<Locale, String> localizedMap =
-							_getLocalizedNamesFromProperties(unicodeProperties);
-
-						Set<Locale> locales = localizedMap.keySet();
-
-						return LocaleUtil.toW3cLanguageIds(
-							locales.toArray(new Locale[localizedMap.size()]));
-					});
-				setCreator(
-					() -> {
-						User user = _userLocalService.fetchUser(
-							siteNavigationMenuItem.getUserId());
-
-						if (user == null) {
-							return null;
-						}
-
-						return new Creator() {
-							{
-								setExternalReferenceCode(
-									user::getExternalReferenceCode);
-							}
-						};
-					});
-				setCustomFields(
-					() -> CustomFieldsUtil.toCustomFields(
-						contextAcceptLanguage.isAcceptAllLanguages(),
-						SiteNavigationMenuItem.class.getName(),
-						siteNavigationMenuItem.getSiteNavigationMenuItemId(),
-						siteNavigationMenuItem.getCompanyId(),
-						contextAcceptLanguage.getPreferredLocale()));
-				setDateCreated(siteNavigationMenuItem::getCreateDate);
-				setDateModified(siteNavigationMenuItem::getModifiedDate);
-				setDefaultLanguageId(
-					() -> unicodeProperties.getProperty("defaultLanguageId"));
-				setDisplayIcon(
-					() -> unicodeProperties.getProperty("displayIcon"));
-				setExternalReferenceCode(
-					siteNavigationMenuItem::getExternalReferenceCode);
-				setId(siteNavigationMenuItem::getSiteNavigationMenuItemId);
-				setName(
-					() -> _getName(
-						layout, navigationMenuItemType, unicodeProperties,
-						getUseCustomName()));
-				setName_i18n(
-					() -> {
-						Map<Locale, String> localizedNames =
-							_getLocalizedNamesFromProperties(unicodeProperties);
-
-						if ((!useCustomName || localizedNames.isEmpty()) &&
-							(layout != null)) {
-
-							localizedNames = layout.getNameMap();
-						}
-
-						return LocalizedMapUtil.getI18nMap(localizedNames);
-					});
-				setNavigationMenuItems(
-					() -> transformToArray(
-						siteNavigationMenuItemsMap.getOrDefault(
-							siteNavigationMenuItem.
-								getSiteNavigationMenuItemId(),
-							new ArrayList<>()),
-						item -> _toNavigationMenuItem(
-							item, siteNavigationMenuItemsMap),
-						NavigationMenuItem.class));
-				setType(siteNavigationMenuItem::getType);
-
-				Object navigationMenuItemSettings =
-					_getNavigationMenuItemSettings(
-						siteNavigationMenuItem.getType(), unicodeProperties);
-
-				if (navigationMenuItemSettings != null) {
-					setNavigationMenuItemSettings(
-						() -> navigationMenuItemSettings);
-				}
-
-				setUseCustomName(
-					() -> Boolean.valueOf(
-						unicodeProperties.getProperty("useCustomName")));
-			}
-		};
-	}
-
-	private Permission[] _toPermissions(SiteNavigationMenu siteNavigationMenu)
+			SiteNavigationMenu siteNavigationMenu)
 		throws Exception {
 
-		return NestedFieldsSupplier.supply(
-			"permissions",
-			nestedFieldNames -> {
-				_permissionService.checkPermission(
-					siteNavigationMenu.getGroupId(),
-					siteNavigationMenu.getModelClassName(),
-					siteNavigationMenu.getSiteNavigationMenuId());
-
-				Collection<Permission> permissions =
-					PermissionUtil.getPermissions(
-						siteNavigationMenu.getCompanyId(),
-						_resourceActionLocalService.getResourceActions(
-							siteNavigationMenu.getModelClassName()),
-						siteNavigationMenu.getSiteNavigationMenuId(),
-						siteNavigationMenu.getModelClassName(), null);
-
-				return permissions.toArray(new Permission[0]);
-			});
-	}
-
-	private String _toType(String type) {
-		if (type.equals("layout")) {
-			return "page";
-		}
-		else if (type.equals("node")) {
-			return "navigationMenu";
-		}
-		else if (type.equals(FileEntry.class.getName())) {
-			return DLFileEntry.class.getName();
-		}
-
-		return type;
+		return _navigationMenuDTOConverter.toDTO(
+			new DefaultDTOConverterContext(
+				contextAcceptLanguage.isAcceptAllLanguages(),
+				HashMapBuilder.put(
+					"delete",
+					addAction(
+						ActionKeys.DELETE, siteNavigationMenu,
+						"deleteSiteNavigationMenu")
+				).put(
+					"replace",
+					addAction(
+						ActionKeys.UPDATE, siteNavigationMenu,
+						"putSiteNavigationMenu")
+				).build(),
+				_dtoConverterRegistry, contextHttpServletRequest,
+				siteNavigationMenu.getSiteNavigationMenuId(),
+				contextAcceptLanguage.getPreferredLocale(), contextUriInfo,
+				contextUser),
+			siteNavigationMenu);
 	}
 
 	private NavigationMenu _updateNavigationMenu(
@@ -1233,19 +777,16 @@ public class NavigationMenuResourceImpl
 		new NavigationMenuEntityModel();
 
 	@Reference
-	private GroupLocalService _groupLocalService;
+	private DTOConverterRegistry _dtoConverterRegistry;
 
 	@Reference
 	private JSONFactory _jsonFactory;
 
-	@Reference
-	private LayoutLocalService _layoutLocalService;
-
-	@Reference
-	private PermissionService _permissionService;
-
-	@Reference
-	private ResourceActionLocalService _resourceActionLocalService;
+	@Reference(
+		target = "(component.name=com.liferay.headless.admin.site.internal.dto.v1_0.converter.NavigationMenuDTOConverter)"
+	)
+	private DTOConverter<SiteNavigationMenu, NavigationMenu>
+		_navigationMenuDTOConverter;
 
 	@Reference
 	private SiteNavigationMenuItemService _siteNavigationMenuItemService;
@@ -1259,8 +800,5 @@ public class NavigationMenuResourceImpl
 
 	@Reference
 	private SiteNavigationMenuService _siteNavigationMenuService;
-
-	@Reference
-	private UserLocalService _userLocalService;
 
 }
